split repo into files
